ui: views: merge all game views into one file
