+++ dnd-bot/dnd_bot/dc/ui/views/view_lobby.py
@@ -2,7 +2,7 @@ import nextcord
 
 from dnd_bot.dc.ui.message_templates import MessageTemplates
 from dnd_bot.dc.ui.messager import Messager
-from dnd_bot.dc.ui.views.view_movement import ViewMovement
+from dnd_bot.dc.ui.views.view_game import ViewMain
 from dnd_bot.dc.utils.utils import get_user_name_by_id
 from dnd_bot.logic.game.handler_game import HandlerGame
 from dnd_bot.logic.lobby.handler_join import HandlerJoin
@@ -89,7 +89,7 @@ class StartButton(nextcord.ui.View):
                     map_view_message = MessageTemplates. \
                         map_view_template(self.token, Multiverse.get_game(self.token).get_active_player().name,
                                           player.action_points, True)
-                    await Messager.send_dm_message(user, map_view_message, view=ViewMovement(self.token))
+                    await Messager.send_dm_message(user, map_view_message, view=ViewMain(self.token))
                 else:
                     map_view_message = MessageTemplates. \
                         map_view_template(self.token, Multiverse.get_game(self.token).get_active_player().name,
